refactor(reporter): move FailingTestsReporter logic to another class
